Start converting www to TypeScript
