+++ typings/index.d.ts
@@ -1,8 +1,6 @@
 // Silence errors from libraries with no type defintions provided
 declare module 'intersection-observer';
 declare module 'object-fit-images';
-// TODO(giles): remove once we add types to tp-icons
-declare module '@thumbtack/thumbprint-icons';
 
 declare module 'rotate-array' {
     function rotate<T>(array: Array<T>, offset: number): Array<T>;
